+++ Common/src/main/java/com/hypherionmc/sdlink/core/config/SDLinkConfig.java
@@ -39,7 +39,7 @@ public final class SDLinkConfig extends AbstractConfig<SDLinkConfig> {
     // DO NOT REMOVE TRANSIENT HERE... OTHERWISE, THE STUPID CONFIG LIBRARY
     // WILL TRY TO WRITE THESE TO THE CONFIG
     public transient static SDLinkConfig INSTANCE;
-    public transient static int configVer = 32;
+    public transient static int configVer = 33;
     public transient static boolean hasConfigLoaded = false;
     public transient static boolean wasReload = false;